+++ Configuration/Extbase/Persistence/Classes.php
@@ -35,6 +35,8 @@ return [
             EventInterface::TYPE_EVENT_DATE => EventDate::class,
         ],
         'properties' => [
+            // This is only needed so we can make queries on the internal title before the actual type is known.
+            'internalTitle' => ['fieldName' => 'title'],
             // This is only needed so we can make queries on the end date before the actual type is known.
             'end' => ['fieldName' => 'end_date'],
             // This is only needed so we can make queries on the start date before the actual type is known.